pyproject.toml
[project]
name = "human-dream-dataset"
version = "0.0.1"
description = "Human Dream Dataset for BCI and Simulation Research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "pytest>=8",
]

[project.optional-dependencies]
bot = [
    "python-telegram-bot>=21",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]
dev = [
    "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "analysis"]
